+++ src/di/src/Aop/AbstractAspect.php
@@ -26,4 +26,9 @@ abstract class AbstractAspect implements AroundInterface
      * @var array
      */
     public $annotations = [];
+
+    /**
+     * @var null|int
+     */
+    public $priority;
 }