+++ src/model-listener/src/Collector/ListenerCollector.php
@@ -52,6 +52,20 @@ class ListenerCollector extends MetadataCollector
      */
     public static function clearListeners(): void
     {
-        static::$container = [];
+        static::clear();
+    }
+
+    public static function clear(?string $listener = null): void
+    {
+        if ($listener) {
+            foreach (static::$container as $model => $listeners) {
+                if ($id = array_search($listener, $listeners)) {
+                    unset($listeners[$id]);
+                    static::$container[$model] = array_values($listeners);
+                }
+            }
+        } else {
+            static::$container = [];
+        }
     }
 }