+++ src/pool/tests/FrequencyTest.php
@@ -11,7 +11,12 @@ declare(strict_types=1);
  */
 namespace HyperfTest\Pool;
 
+use Hyperf\Pool\Channel;
+use Hyperf\Pool\Pool;
+use Hyperf\Utils\Coroutine;
+use HyperfTest\Pool\Stub\ConstantFrequencyStub;
 use HyperfTest\Pool\Stub\FrequencyStub;
+use Mockery;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -20,6 +25,11 @@ use PHPUnit\Framework\TestCase;
  */
 class FrequencyTest extends TestCase
 {
+    protected function tearDown()
+    {
+        Mockery::close();
+    }
+
     public function testFrequencyHit()
     {
         $frequency = new FrequencyStub();
@@ -39,6 +49,19 @@ class FrequencyTest extends TestCase
         $frequency->hit();
         $num = $frequency->frequency();
         $this->assertSame(42 / 5, $num);
+    }
+
+    public function testConstantFrequency()
+    {
+        $pool = Mockery::mock(Pool::class);
+        $channel = new Channel(100);
+        $pool->shouldReceive('flushOne')->andReturnUsing(function () use ($channel) {
+            $channel->push(true);
+        });
+
+        new ConstantFrequencyStub($pool);
+        Coroutine::sleep(0.01);
+        $this->assertGreaterThan(0, $channel->length());
     }
 
     public function testFrequencyHitOneSecondAfter()