+++ src/rpc-client/src/Proxy/CodeLoader.php
@@ -23,4 +23,9 @@ class CodeLoader
         }
         return file_get_contents($file);
     }
+
+    public function getPathByClassName(string $className): string
+    {
+        return Composer::getLoader()->findFile($className);
+    }
 }