+++ src/rpc-server/src/RequestDispatcher.php
@@ -16,6 +16,7 @@ use Hyperf\Dispatcher\HttpRequestHandler;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
 
 class RequestDispatcher extends HttpDispatcher
 {
@@ -29,14 +30,11 @@ class RequestDispatcher extends HttpDispatcher
         $this->container = $container;
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function dispatch(...$params): ResponseInterface
     {
         /** @var ServerRequestInterface $request */
         /** @var array $middlewares */
-        /** @var string $coreHandler */
+        /** @var MiddlewareInterface $coreHandler */
         [$request, $middlewares, $coreHandler] = $params;
         $requestHandler = new HttpRequestHandler($middlewares, $coreHandler, $this->container);
         return $requestHandler->handle($request);