+++ src/utils/tests/Codec/JsonTest.php
@@ -47,17 +47,13 @@ class JsonTest extends TestCase
         $this->assertEquals((object) $data, Json::decode($json, false));
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     * @expectedExceptionMessage Control character error, possibly incorrectly encoded
-     */
     public function testDecodeException()
     {
-        $data = [
-            'name' => 'Hyperf',
-        ];
         $json = '{"name":"Hyperf}';
-        $this->assertSame($data, Json::decode($json));
+
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('Control character error, possibly incorrectly encoded');
+        Json::decode($json);
     }
 
     public function testJsonEncodeInCoroutine()